refactor(playground): move file
