+++ src/playground/editor/code/data/coc7/coc7-fund.js
