+++ src/playground/editor/code/data/dnd5e/README.md
