+++ src/playground/editor/code/data/dnd5e/classes.js
