+++ src/playground/editor/code/data/dnd5e/level.js
