+++ src/playground/editor/code/data/dnd5e/weapon.js
