+++ src/playground/editor/code/example/xml/coc7.xml
