+++ src/playground/editor/code/example/xml/display.xml
