+++ src/playground/editor/code/example/xml/dnd5e.xml
