+++ src/playground/editor/code/example/xml/function.xml
