+++ src/playground/editor/code/example/xml/input.xml
