+++ src/playground/editor/code/example/xml/simple.xml
