+++ src/playground/editor/code/example/xml/tabs.xml
