+++ src/playground/editor/code/example/xml/use-define.xml
