+++ src/playground/editor/code/example/xml/var.xml
