+++ src/playground/editor/code/example/xml/wuxian-simple.xml
