+++ src/playground/editor/code/example/xml/wuxian.xml
